requirements.txt
pandas>=2.1.0
openpyxl>=3.1.2
streamlit>=1.32.0
openai>=1.12.0
langchain 
langchain-openai
python-dotenv
pydantic>=2.5.0